Use `Long` by default in the template

+++ .idea/fileTemplates/Advent of Code.kt
@@ -1,9 +1,9 @@
 fun main() {
-    fun part1(input: List<String>): Int {
+    fun part1(input: List<String>): Long {
         return TODO()
     }
 
-    fun part2(input: List<String>): Int {
+    fun part2(input: List<String>): Long {
         return TODO()
     }
